+++ classes/BlueChip/Security/Modules/Log/Logger.php
@@ -10,7 +10,6 @@ use BlueChip\Security\Modules\Cron\Jobs as CronJobs;
 use BlueChip\Security\Modules\Services\ReverseDnsLookup\Resolver;
 use BlueChip\Security\Modules\Services\ReverseDnsLookup\Response;
 use Psr\Log\AbstractLogger;
-use Psr\Log\LoggerInterface;
 use Psr\Log\LogLevel;
 use wpdb;
 
@@ -19,7 +18,7 @@ use wpdb;
  *
  * @link http://www.php-fig.org/psr/psr-3/
  */
-class Logger extends AbstractLogger implements LoggerInterface, Modules\Countable, Modules\Installable, Modules\Loadable, Modules\Initializable, \Countable
+class Logger extends AbstractLogger implements Modules\Countable, Modules\Installable, Modules\Loadable, Modules\Initializable, \Countable
 {
     /**
      * @var string Name of DB table where logs are stored
@@ -59,7 +58,7 @@ class Logger extends AbstractLogger implements LoggerInterface, Modules\Countabl
     public function install(): void
     {
         // To have dbDelta()
-        require_once ABSPATH . 'wp-admin/includes/upgrade.php'; // @phpstan-ignore-line
+        require_once ABSPATH . 'wp-admin/includes/upgrade.php';
 
         $charset_collate = $this->wpdb->get_charset_collate();
 
@@ -117,10 +116,10 @@ class Logger extends AbstractLogger implements LoggerInterface, Modules\Countabl
      * Log generic event.
      *
      * @param string $level
-     * @param string $message
+     * @param string|\Stringable $message
      * @param array<string,mixed> $context
      */
-    public function log($level, $message, array $context = [])
+    public function log($level, string|\Stringable $message, array $context = []): void
     {
         // Allow overriding of IP address via $context.
         $ip_address = $context['ip_address'] ?? $this->remote_address;
@@ -274,7 +273,8 @@ class Logger extends AbstractLogger implements LoggerInterface, Modules\Countabl
     {
         /** @var string $query */
         $query = $this->wpdb->prepare(
-            "SELECT COUNT(id) AS total FROM {$this->log_table} WHERE date_and_time > %s",
+            'SELECT COUNT(id) AS total FROM %i WHERE date_and_time > %s',
+            $this->log_table,
             MySQLDateTime::formatDateTime($timestamp)
         );
 
@@ -332,7 +332,11 @@ class Logger extends AbstractLogger implements LoggerInterface, Modules\Countabl
     public function getKnownIps(): array
     {
         /** @var string $query */
-        $query = $this->wpdb->prepare("SELECT DISTINCT(ip_address) FROM {$this->log_table} WHERE event = %s", Events\LoginSuccessful::ID);
+        $query = $this->wpdb->prepare(
+            'SELECT DISTINCT(ip_address) FROM %i WHERE event = %s',
+            $this->log_table,
+            Events\LoginSuccessful::ID
+        );
 
         $result = $this->wpdb->get_results($query, ARRAY_A);
 
@@ -363,7 +367,8 @@ class Logger extends AbstractLogger implements LoggerInterface, Modules\Countabl
         // Note: $wpdb->delete cannot be used as it does not support "<=" comparison)
         /** @var string $query */
         $query = $this->wpdb->prepare(
-            "DELETE FROM {$this->log_table} WHERE date_and_time <= %s",
+            'DELETE FROM %i WHERE date_and_time <= %s',
+            $this->log_table,
             MySQLDateTime::formatDateTime(\time() - $max_age)
         );
         // Execute query and return true/false status.
@@ -398,14 +403,18 @@ class Logger extends AbstractLogger implements LoggerInterface, Modules\Countabl
 
         // Find the biggest ID from all records that should be pruned.
         /** @var string $query_id */
-        $query_id = $this->wpdb->prepare("SELECT id FROM {$this->log_table} ORDER BY id DESC LIMIT %d, 1", $max_size);
+        $query_id = $this->wpdb->prepare(
+            'SELECT id FROM %i ORDER BY id DESC LIMIT %d, 1',
+            $this->log_table,
+            $max_size
+        );
         if (empty($id = (int) $this->wpdb->get_var($query_id))) {
             return false;
         }
 
         // Note: $wpdb->delete cannot be used as it does not support "<=" comparison)
         /** @var string $query */
-        $query = $this->wpdb->prepare("DELETE FROM {$this->log_table} WHERE id <= %d", $id);
+        $query = $this->wpdb->prepare('DELETE FROM %i WHERE id <= %d', $this->log_table, $id);
         // Execute query and return true/false status.
         return $this->wpdb->query($query) !== false;
     }