+++ classes/Psr/Log/LoggerInterface.php
@@ -24,12 +24,9 @@ interface LoggerInterface
     /**
      * System is unusable.
      *
-     * @param string $message
-     * @param array  $context
-     *
-     * @return void
+     * @param mixed[] $context
      */
-    public function emergency($message, array $context = []);
+    public function emergency(string|\Stringable $message, array $context = []): void;
 
     /**
      * Action must be taken immediately.
@@ -37,35 +34,26 @@ interface LoggerInterface
      * Example: Entire website down, database unavailable, etc. This should
      * trigger the SMS alerts and wake you up.
      *
-     * @param string $message
-     * @param array  $context
-     *
-     * @return void
+     * @param mixed[] $context
      */
-    public function alert($message, array $context = []);
+    public function alert(string|\Stringable $message, array $context = []): void;
 
     /**
      * Critical conditions.
      *
      * Example: Application component unavailable, unexpected exception.
      *
-     * @param string $message
-     * @param array  $context
-     *
-     * @return void
+     * @param mixed[] $context
      */
-    public function critical($message, array $context = []);
+    public function critical(string|\Stringable $message, array $context = []): void;
 
     /**
      * Runtime errors that do not require immediate action but should typically
      * be logged and monitored.
      *
-     * @param string $message
-     * @param array  $context
-     *
-     * @return void
+     * @param mixed[] $context
      */
-    public function error($message, array $context = []);
+    public function error(string|\Stringable $message, array $context = []): void;
 
     /**
      * Exceptional occurrences that are not errors.
@@ -73,53 +61,40 @@ interface LoggerInterface
      * Example: Use of deprecated APIs, poor use of an API, undesirable things
      * that are not necessarily wrong.
      *
-     * @param string $message
-     * @param array  $context
-     *
-     * @return void
+     * @param mixed[] $context
      */
-    public function warning($message, array $context = []);
+    public function warning(string|\Stringable $message, array $context = []): void;
 
     /**
      * Normal but significant events.
      *
-     * @param string $message
-     * @param array  $context
-     *
-     * @return void
+     * @param mixed[] $context
      */
-    public function notice($message, array $context = []);
+    public function notice(string|\Stringable $message, array $context = []): void;
 
     /**
      * Interesting events.
      *
      * Example: User logs in, SQL logs.
      *
-     * @param string $message
-     * @param array  $context
-     *
-     * @return void
+     * @param mixed[] $context
      */
-    public function info($message, array $context = []);
+    public function info(string|\Stringable $message, array $context = []): void;
 
     /**
      * Detailed debug information.
      *
-     * @param string $message
-     * @param array  $context
-     *
-     * @return void
+     * @param mixed[] $context
      */
-    public function debug($message, array $context = []);
+    public function debug(string|\Stringable $message, array $context = []): void;
 
     /**
      * Logs with an arbitrary level.
      *
-     * @param mixed  $level
-     * @param string $message
-     * @param array  $context
+     * @param mixed $level
+     * @param mixed[] $context
      *
-     * @return void
+     * @throws \Psr\Log\InvalidArgumentException
      */
-    public function log($level, $message, array $context = []);
+    public function log($level, string|\Stringable $message, array $context = []): void;
 }